+++ fuzz/fuzzers/fragment.rs
@@ -6,11 +6,11 @@ extern crate rustls;
 use rustls::internal::msgs::base::Payload;
 use rustls::internal::msgs::codec::Reader;
 use rustls::internal::msgs::fragmenter::MessageFragmenter;
-use rustls::internal::msgs::message::{Message, OpaqueMessage, PlainMessage};
+use rustls::internal::msgs::message::{Message, OutboundOpaqueMessage, PlainMessage};
 
 fuzz_target!(|data: &[u8]| {
     let mut rdr = Reader::init(data);
-    let msg = match OpaqueMessage::read(&mut rdr) {
+    let msg = match OutboundOpaqueMessage::read(&mut rdr) {
         Ok(msg) => msg,
         Err(_) => return,
     };
