+++ fuzz/fuzzers/message.rs
@@ -4,11 +4,11 @@ extern crate libfuzzer_sys;
 extern crate rustls;
 
 use rustls::internal::msgs::codec::Reader;
-use rustls::internal::msgs::message::{Message, OpaqueMessage, PlainMessage};
+use rustls::internal::msgs::message::{Message, OutboundOpaqueMessage, PlainMessage};
 
 fuzz_target!(|data: &[u8]| {
     let mut rdr = Reader::init(data);
-    if let Ok(m) = OpaqueMessage::read(&mut rdr) {
+    if let Ok(m) = OutboundOpaqueMessage::read(&mut rdr) {
         let msg = match Message::try_from(m.into_plain_message()) {
             Ok(msg) => msg,
             Err(_) => return,
