+++ rustls/src/msgs/base.rs
@@ -4,12 +4,9 @@ use crate::msgs::codec::{Codec, Reader};
 
 use alloc::vec::Vec;
 use core::fmt;
-use core::ops::{Deref, DerefMut};
 
 use pki_types::CertificateDer;
 use zeroize::Zeroize;
-
-use super::codec::ReaderMut;
 
 /// An externally length'd payload
 #[derive(Clone, Eq, PartialEq)]
@@ -59,59 +56,6 @@ impl Payload<'static> {
 
     pub fn empty() -> Self {
         Self::Borrowed(&[])
-    }
-}
-
-/// Non-owning version of [`Payload`]
-pub struct BorrowedPayload<'a>(&'a mut [u8]);
-
-impl Deref for BorrowedPayload<'_> {
-    type Target = [u8];
-
-    fn deref(&self) -> &Self::Target {
-        self.0
-    }
-}
-
-impl<'a> DerefMut for BorrowedPayload<'a> {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        self.0
-    }
-}
-
-impl<'a> BorrowedPayload<'a> {
-    #[cfg(test)]
-    pub(crate) fn new(bytes: &'a mut [u8]) -> Self {
-        Self(bytes)
-    }
-
-    pub fn truncate(&mut self, len: usize) {
-        if len >= self.len() {
-            return;
-        }
-
-        self.0 = core::mem::take(&mut self.0)
-            .split_at_mut(len)
-            .0;
-    }
-
-    pub(crate) fn read(r: &mut ReaderMut<'a>) -> Self {
-        Self(r.rest())
-    }
-
-    pub(crate) fn into_inner(self) -> &'a mut [u8] {
-        self.0
-    }
-
-    pub(crate) fn pop(&mut self) -> Option<u8> {
-        if self.is_empty() {
-            return None;
-        }
-
-        let len = self.len();
-        let last = self[len - 1];
-        self.truncate(len - 1);
-        Some(last)
     }
 }
