+++ rustls/src/msgs/message_test.rs
@@ -4,7 +4,7 @@ use crate::msgs::base::{PayloadU16, PayloadU24, PayloadU8};
 use super::base::Payload;
 use super::codec::Reader;
 use super::enums::AlertLevel;
-use super::message::{Message, OpaqueMessage, PlainMessage};
+use super::message::{Message, OutboundOpaqueMessage, PlainMessage};
 
 use std::fs;
 use std::io::Read;
@@ -29,7 +29,7 @@ fn test_read_fuzz_corpus() {
         f.read_to_end(&mut bytes).unwrap();
 
         let mut rd = Reader::init(&bytes);
-        let msg = OpaqueMessage::read(&mut rd)
+        let msg = OutboundOpaqueMessage::read(&mut rd)
             .unwrap()
             .into_plain_message();
         println!("{:?}", msg);
@@ -70,7 +70,7 @@ fn can_read_safari_client_hello() {
         \x79\x2f\x33\x08\x68\x74\x74\x70\x2f\x31\x2e\x31\x00\x0b\x00\x02\
         \x01\x00\x00\x0a\x00\x0a\x00\x08\x00\x1d\x00\x17\x00\x18\x00\x19";
     let mut rd = Reader::init(bytes);
-    let m = OpaqueMessage::read(&mut rd).unwrap();
+    let m = OutboundOpaqueMessage::read(&mut rd).unwrap();
     println!("m = {:?}", m);
     assert!(Message::try_from(m.into_plain_message()).is_err());
 }
@@ -91,7 +91,7 @@ fn construct_all_types() {
         &b"\x18\x03\x04\x00\x04\x11\x22\x33\x44"[..],
     ];
     for &bytes in samples.iter() {
-        let m = OpaqueMessage::read(&mut Reader::init(bytes)).unwrap();
+        let m = OutboundOpaqueMessage::read(&mut Reader::init(bytes)).unwrap();
         println!("m = {:?}", m);
         let m = Message::try_from(m.into_plain_message());
         println!("m' = {:?}", m);
