+++ rustls/src/msgs/mod.rs
@@ -25,7 +25,7 @@ mod message_test;
 #[cfg(test)]
 mod tests {
     use super::codec::Reader;
-    use super::message::{Message, OpaqueMessage};
+    use super::message::{Message, OutboundOpaqueMessage};
 
     #[test]
     fn smoketest() {
@@ -33,7 +33,7 @@ mod tests {
         let mut r = Reader::init(bytes);
 
         while r.any_left() {
-            let m = OpaqueMessage::read(&mut r).unwrap();
+            let m = OutboundOpaqueMessage::read(&mut r).unwrap();
 
             let out = m.clone().encode();
             assert!(!out.is_empty());
